+++ Code/ForceField/UFF/AngleBend.h
@@ -38,9 +38,9 @@ class RDKIT_FORCEFIELD_EXPORT AngleBendContrib : public ForceFieldContrib {
     \param order (optional) the order of the angle term, this is for
        special cases and should adopt values:
          - 0: not a special case, use the \c theta0 value from \c at2Params
-         - 2: linear coordination
+         - 1: linear coordination
          - 3: trigonal planar coordination
-         - 4: square planar or tetrahedral coordination
+         - 4: square planar or octahedral coordination
 
   */
   AngleBendContrib(ForceField *owner, unsigned int idx1, unsigned int idx2,